Add SRID, and WKB marshalling for `Point`
The former is 25 bytes while the latter is 21 bytes.

The marshalling to SRID is done using little endian
as MariaDB struggles to handle big endian (at least
without any additional configuration). It is 'left padded'
with 4 zero bytes in the front.

The marshalling is done in `Value`, an implementation of the
`Valuer` interface in `database/sql/driver` which
will allow us to perform writes to the database
when using `Point` as a data type in a struct.

There is also a `marshalXY(..)` method to act as a
parallel for `unmarshalXY(..)`.

The endianness for WKB marshalling can be configured.

The header size can now be set using constants rather
than scattered hardcoding.

See:

- https://mariadb.com/kb/en/library/st_srid/
- https://mariadb.com/kb/en/library/well-known-binary-wkb-format/